Fix handling of zero-size arrays
Before, if `m > 0 && n > 0 && k == 0`, the result was never
initialized or calculated. Additionally, in some other cases, pointer
offsets could result in undefined behavior due to violating the
constraints on the `.offset()` method. (For example, consider the case
where `m == 0 && n > 0` and the `c` pointer is a null pointer (which
is valid because the `c` matrix is empty). Then, the call
`c.stride_offset(csc, knc * l5)` was undefined behavior.)