FEAT: Schedule loads an iteration ahead in avx kernel
This improves performance by a massive 10% on the 127×127 benchmark.

We can assume the kernel is called with k > 0, so we split off the last
iteration of the loop and make sure we don't load in the last iteration
(as it would go out of bounds of the packing area).

There was an alternative, we could extend the packing buffer by 8 dummy
elements to allow us to load up to one more vector, but it did not make
a difference in benchmarks.