TEST: Change test environment variable to use MMTEST_FEATURE=avx etc
Instead of excluding features to test, *include* them.

If you set MMTEST_FEATURE=avx, only the "avx" feature ever tests true at
runtime. We include "TEST" in the environment variable name so that it
is clear. "MM" is for matrixmultiply.

Also add a test that can be optionally enabled, to ensure we test
a feature.