FIX: Use no inline on gemm_loop
For the default case (runtime feature detect), this could be called with
many different kernels, but only one kernel is ever used on a particular
configuration.

This way the unused kernels end up in separate functions, so the unused
gemm_loop functions are kept separated like this when they are not all
inlined into the top level sgemm/dgemm function.