FIX: Aliasing issue in scale add
It would compile to code here that fetched the value in *cptr twice,
because the compiler can't figure out that cptr and ab (and pointers
derived from them respectiely) never alias.

Work around this by separating into another function that takes the
mask_buf as a &T argument (which is noalias).