complex: compute cgemm, zgemm in real parts
Combine the cgemm, zgemm kernels.  Break out components
and compute the real/imag parts separately. This has especially large
gains for FMA.

A separate approach was also tried, using (P + Qi)(R + Si) and
factoring into three separate multiplications (instead of four),
and this was a benefit for cgemm, but not as big a benefit as using
the simpler kernel in this PR, compiled using FMA.