tests: Combine repeated generic code in tests and benchmark
include!() was chosen for sharing this code. An internal crate could
have been used, but it has the downside that the interal crate doesn't
become part of the source package published on crates.io, which is
nonideal (just a minor thing, but still).

In the spirit of open source, the source package should contain the
preferred setup for working with the project, and that includes the
tests. With this solution, the tests are still buildable from the
published package.