test: Factor out common matrix compare
The new matrix compare looks at equality within a tolerance.

As-written, the testsuite would end up with only exact integer floats,
or additions where it loses precisions in the expected way.

When changing KC to a smaller value or, when testing larger matrices,
because this impacts loop 4 and we do one update (+=) to C per iteration
of loop 4 - then this can be visible in the precision or rounding error
of the result. Thus, with varying KC - we must have a relative tolerance for
equality.