ptr: Silence suspicious Send/Sync impls warning
Revisit the previous fix. It's clear that the Ptr code will not change
meaning with the coming breaking change, because it's only ever used
with Ptr<*const T> and Ptr<*mut T> as parameterizations.

For this reason, it's more correct IMO to accept the code meaning
change, which doesn't change the meaning of the crate, by silencing the
warning and not making any unnecessary and ugly changes to the struct
fields.