x86-64: Specialize pack function for avx2
Because we detect target features to select kernel, and the kernel
can select its own packing functions, we can now specialize the packing
functions per target.

As matrices get larger, the packing performance matters much less, but
for small matrix products it contributes more to the runtime.

The default packing also already has a special case for contiguous
matrices, which happens when in C = A B, A is column major and B is row
major. The specialization in this commit helps the most outside this
special case.