docs: Add document retracing x86-64 AVX sgemm microkernel in typst
This is mostly for fun (and verification). Not as generic as it could be
and so on.

Compiled using typst 0.13.1. PDF included in repo so that it is readily
available to read.

Experimenting with the document in typst.app or locally with instant
preview is a good way to work with it.